+++ apps/CASTool_USEPA/global.R
@@ -5,7 +5,7 @@
 #~~~~~~~~~~~~~~~~~~~~~~~~
 
 # Version ----
-pkg_version <- "0.0.2.9009"
+pkg_version <- "0.0.2.9010"
 
 # Packages ----
 library(shiny)